+++ example/bug-fixed/local-repo-config/xperc/perc/perc.cc
@@ -1,0 +1,11 @@
+#include "perc/perc.h"
+
+namespace xmake {
+namespace perc {
+
+int add(const int a, const int b) {
+    return a + b;
+}
+
+}
+}