+++ example/bug-fixed/local-repo-config/xsim/main.cc
@@ -1,0 +1,7 @@
+#include <iostream>
+#include "perc/perc.h"
+
+int main() {
+    int sum = xmake::perc::add(10, 20);
+    std::cout << "sum: " << sum << std::endl;
+}